% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{about_card}
\alias{about_card}
\alias{guides_card}
\alias{sources_card}
\title{Static cards for the "More" sidebar panel}
\usage{
about_card()

guides_card()

sources_card()
}
\description{
Static cards for the "More" sidebar panel
}
\keyword{internal}
