% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{complete_usage_gaps_with_zeros}
\alias{complete_usage_gaps_with_zeros}
\title{Helper function fill usage for missing years}
\usage{
complete_usage_gaps_with_zeros(data)
}
\description{
This only fills gaps between existing start and end dates for each code
but does not extent the date range for a code.
}
\keyword{internal}
