% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-sources.R
\name{download_tidy_source}
\alias{download_tidy_source}
\title{Download a dataset parquet and store it in the cache}
\usage{
download_tidy_source(dataset, url, version)
}
\arguments{
\item{dataset}{String, dataset name}

\item{url}{String, download URL}

\item{version}{String, dataset version (included in the cached filename)}
}
\value{
Invisible path to the cached parquet
}
\description{
Download a dataset parquet and store it in the cache
}
\keyword{internal}
