% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_codes_table.R
\name{empty_selected_codes}
\alias{empty_selected_codes}
\title{Empty selected codes table}
\usage{
empty_selected_codes()
}
\description{
Used for both the displayed table and the CSV download when no codes or
codelist are selected.
}
\keyword{internal}
