% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{end_date_breaks}
\alias{end_date_breaks}
\title{Pick x-axis breaks from the yearly reporting end dates}
\usage{
end_date_breaks(end_dates, n_max = 6)
}
\description{
Breaks always fall on actual end dates of the yearly reporting periods.
When there are more than \code{n_max}, only every k-th end date is shown,
anchored on the most recent period so the latest data is always labelled.
}
\keyword{internal}
