% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{fill_missing_usage_with_zeros}
\alias{fill_missing_usage_with_zeros}
\title{Helper function to fill 0 in missing years for plotting individual code usage}
\usage{
fill_missing_usage_with_zeros(data)
}
\description{
Helper function to fill 0 in missing years for plotting individual code usage
}
\keyword{internal}
