% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opencodelists.R
\name{get_codelist}
\alias{get_codelist}
\title{Get codelist from \href{https://www.opencodelists.org}{OpenCodelists}}
\usage{
get_codelist(url)
}
\arguments{
\item{url}{String, specifying URL to codelist on \href{https://www.opencodelists.org}{OpenCodelists}}
}
\description{
Get codelist from \href{https://www.opencodelists.org}{OpenCodelists}
}
\examples{
# Get the 'cpeptide_cod' codelist from OpenCodelists.org
codelist_url <- paste0(
  "https://www.opencodelists.org/codelist/",
  "nhsd-primary-care-domain-refsets/cpeptide_cod/20200812/"
)
cpeptide_cod <- get_codelist(codelist_url)

# Return all codes
cpeptide_cod$code

# Return 'coding_system' of codelist
cpeptide_cod@coding_system

# Return 'full_slug' of codelist
cpeptide_cod@full_slug
}
