% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_codes_with_encoding_problems}
\alias{get_codes_with_encoding_problems}
\title{Find codes with encoding problems}
\usage{
get_codes_with_encoding_problems(code_usage_data, code_column)
}
\description{
Find codes with encoding problems
}
\keyword{internal}
