% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_codes_with_multiple_desc}
\alias{get_codes_with_multiple_desc}
\title{Find codes with multiple description}
\usage{
get_codes_with_multiple_desc(code_usage_data)
}
\description{
Find codes with multiple description
}
\keyword{internal}
