% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_data_cache_dir}
\alias{get_data_cache_dir}
\title{Get data cache directory}
\usage{
get_data_cache_dir()
}
\value{
Character path to the directory holding cached dataset parquet files
}
\description{
Get data cache directory
}
\keyword{internal}
