% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_dataset}
\alias{get_dataset}
\title{Get a dataset, downloading and caching it if necessary}
\usage{
get_dataset(dataset, version = NULL)
}
\arguments{
\item{dataset}{String, dataset name as listed in \code{tidy_data_sources.yml}}

\item{version}{String, dataset version, or \code{NULL} for the latest}
}
\value{
Tibble
}
\description{
Get a dataset, downloading and caching it if necessary
}
\keyword{internal}
