% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_hesapc_icd10}
\alias{get_hesapc_icd10}
\title{Get yearly ICD-10 diagnosis code usage from Hospital Episode Statistics Admitted Patient Care in England}
\source{
\url{https://digital.nhs.uk/data-and-information/publications/statistical/hospital-admitted-patient-care-activity}
}
\usage{
get_hesapc_icd10(version = NULL)
}
\arguments{
\item{version}{String, dataset version, or \code{NULL} for the latest (default)}
}
\value{
A tibble
}
\description{
Downloads (if not already cached) and returns the yearly summary of
4-character ICD-10 code usage from 1st April 2012 onwards. The code usage
represents the annual count of all episodes which record the given ICD-10
code in any primary or secondary position. Restricted codes for which
annual usage is not published have been removed.
\describe{
\item{start_date}{Start date of code usage count}
\item{end_date}{End date of code usage count}
\item{icd10_code}{The 4-character ICD-10 Code.
Note that the punctuation from the code has been removed for compatibility with OpenCodelists.}
\item{usage}{Annual count of code usage.}
\item{description}{Description of the ICD-10 Code}
}
}
\examples{
\dontrun{
# Filter to codes in the ICD-10 Chapter XIX: "Injury, poisoning..."
# (codes begin with letters "S" or "T"), with usage > 10,000.
get_hesapc_icd10() |>
  dplyr::filter(grepl("^[ST]", icd10_code) & usage > 10000)
}
}
