% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_opcs4_usage}
\alias{get_opcs4_usage}
\title{Get yearly OPCS-4 code usage from Hospital Admitted Patient Care Activity in England}
\source{
\url{https://digital.nhs.uk/data-and-information/publications/statistical/hospital-admitted-patient-care-activity}
}
\usage{
get_opcs4_usage(version = NULL)
}
\arguments{
\item{version}{String, dataset version, or \code{NULL} for the latest (default)}
}
\value{
A tibble
}
\description{
Downloads (if not already cached) and returns the yearly summary of
4-character OPCS-4 code usage from 1st April 2012 onwards. The code usage
represents the total annual count of each procedure, recorded across the
primary and the secondary procedure positions. Restricted codes for which
annual usage is not published have been removed.
\describe{
\item{start_date}{Start date of code usage count}
\item{end_date}{End date of code usage count}
\item{opcs4_code}{The 4-character OPCS-4 code.
Note that the punctuation from the code has been removed for compatibility with OpenCodelists.}
\item{usage}{Annual count of code usage.}
\item{description}{Description of the OPCS-4 Code}
}
}
\examples{
\dontrun{
# Filter to procedures involving "biopsy" after March 2020 (note each year runs April - March).
get_opcs4_usage() |>
  dplyr::filter(grepl("biopsy", description, ignore.case = TRUE) & lubridate::year(end_date) > 2020)
}
}
