% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_snomed_usage}
\alias{get_snomed_usage}
\title{Get yearly SNOMED CT code usage in primary care in England}
\usage{
get_snomed_usage(version = NULL)
}
\arguments{
\item{version}{String, dataset version, or \code{NULL} for the latest (default)}
}
\value{
A tibble
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
Renamed to \code{\link[=get_gp_snomed]{get_gp_snomed()}}.
}
\keyword{internal}
