% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-sources.R
\name{get_tidy_source_cache_path}
\alias{get_tidy_source_cache_path}
\title{Get path to cached dataset parquet}
\usage{
get_tidy_source_cache_path(dataset, version)
}
\arguments{
\item{dataset}{String, dataset name}

\item{version}{String, dataset version}
}
\value{
Character path
}
\description{
Get path to cached dataset parquet
}
\keyword{internal}
