% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-sources.R
\name{load_tidy_source}
\alias{load_tidy_source}
\title{Load a dataset parquet from the cache}
\usage{
load_tidy_source(dataset, version)
}
\arguments{
\item{dataset}{String, dataset name}

\item{version}{String, dataset version}
}
\value{
Tibble
}
\description{
Load a dataset parquet from the cache
}
\keyword{internal}
