% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_icd10_usage_breakdowns}
\alias{get_icd10_usage_breakdowns}
\title{Get yearly ICD-10 code usage breakdowns from Hospital Admitted Patient Care Activity in England}
\usage{
get_icd10_usage_breakdowns(version = NULL)
}
\arguments{
\item{version}{String, dataset version, or \code{NULL} for the latest (default)}
}
\value{
A tibble
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
Renamed to \code{\link[=get_hesapc_icd10_breakdowns]{get_hesapc_icd10_breakdowns()}}.
}
\keyword{internal}
