% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_opcs4_usage}
\alias{get_opcs4_usage}
\title{Get yearly OPCS-4 code usage from Hospital Admitted Patient Care Activity in England}
\usage{
get_opcs4_usage(version = NULL)
}
\arguments{
\item{version}{String, dataset version, or \code{NULL} for the latest (default)}
}
\value{
A tibble
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
Renamed to \code{\link[=get_hesapc_opcs4]{get_hesapc_opcs4()}}.
}
\keyword{internal}
