% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_snomed_usage}
\alias{get_snomed_usage}
\title{Get yearly SNOMED CT code usage in primary care in England}
\source{
\url{https://digital.nhs.uk/data-and-information/publications/statistical/mi-snomed-code-usage-in-primary-care}
}
\usage{
get_snomed_usage(version = NULL)
}
\arguments{
\item{version}{String, dataset version, or \code{NULL} for the latest (default)}
}
\value{
A tibble
}
\description{
Downloads (if not already cached) and returns the yearly summary of
SNOMED CT code usage from 1st August 2011 onwards. The variables in this
dataset include:
\describe{
\item{start_date}{Start date of code usage count}
\item{end_date}{End date of code usage count}
\item{snomed_code}{SNOMED Concept ID}
\item{usage}{Yearly summary of code usage.
Note that counts are rounded to the nearest 10.
Counts of 5 or below are displayed as 5.}
\item{active_at_start}{Specifying whether code was active at the start date.}
\item{active_at_end}{Specifying whether code was active at the end date.}
\item{description}{Description of SNOMED Concept ID}
}
}
\examples{
\dontrun{
# Filter for code usage records from 2022-08-01 onwards
get_snomed_usage() |>
  dplyr::filter(start_date >= "2022-08-1")
}
}
